+++ src/arkreactor/Compiler/AST/BaseParser.cpp
@@ -1,9 +1,31 @@
 #include <Ark/Compiler/AST/BaseParser.hpp>
 #include <Ark/Exceptions.hpp>
+
+#include <algorithm>
 #include <utility>
 
 namespace Ark::internal
 {
+    void BaseParser::registerNewLine(std::string::iterator it, std::size_t row)
+    {
+        // search for an existing new line position
+        if (std::find_if(m_it_to_row.begin(), m_it_to_row.end(), [it](const auto& pair) {
+                return pair.first == it;
+            }) != m_it_to_row.end())
+            return;
+
+        for (std::size_t i = 0, end = m_it_to_row.size(); i < end; ++i)
+        {
+            auto current = m_it_to_row[i].first;
+            auto next = i + 1 < end ? m_it_to_row[i + 1].first : m_str.end();
+            if (current < it && it < next)
+            {
+                m_it_to_row.insert(m_it_to_row.begin() + i + 1, std::make_pair(it, row));
+                break;
+            }
+        }
+    }
+
     void BaseParser::next()
     {
         m_it = m_next_it;
@@ -17,6 +39,15 @@ namespace Ark::internal
         auto [it, sym] = utf8_char_t::at(m_it);
         m_next_it = it;
         m_sym = sym;
+
+        if (*m_it == '\n')
+        {
+            ++m_filepos.row;
+            m_filepos.col = 0;
+            registerNewLine(m_it, m_filepos.row);
+        }
+        else if (m_sym.isPrintable())
+            m_filepos.col += m_sym.size();
     }
 
     void BaseParser::initParser(const std::string& filename, const std::string& code)
@@ -46,31 +77,32 @@ namespace Ark::internal
         auto [it, sym] = utf8_char_t::at(m_it);
         m_next_it = it;
         m_sym = sym;
+
+        // TODO: create a kind of map vec<pair<it, row>>
+        // search for the nearest it < m_it in the map to know the line number
+        for (std::size_t i = 0, end = m_it_to_row.size(); i < end; ++i)
+        {
+            auto [at, line] = m_it_to_row[i];
+            if (it < at)
+            {
+                m_filepos.row = line - 1;
+                break;
+            }
+        }
+        // compute the position in the line
+        std::string_view view = m_str;
+        auto it_pos = std::distance(m_str.begin(), m_it);
+        view = view.substr(0, it_pos);
+        auto nearest_newline_index = view.find_last_of('\n');
+        if (nearest_newline_index != std::string_view::npos)
+            m_filepos.col = it_pos - nearest_newline_index + 1;
+        else
+            m_filepos.col = it_pos + 1;
     }
 
     FilePosition BaseParser::getCursor()
     {
-        FilePosition pos { 0, 0 };
-
-        // adjust the row/col count (this is going to be VERY inefficient)
-        auto tmp = m_str.begin();
-        while (true)
-        {
-            auto [it, sym] = utf8_char_t::at(tmp);
-            if (*tmp == '\n')
-            {
-                ++pos.row;
-                pos.col = 0;
-            }
-            else if (sym.isPrintable())
-                pos.col += sym.size();
-            tmp = it;
-
-            if (tmp > m_it || tmp == m_str.end())
-                break;
-        }
-
-        return pos;
+        return m_filepos;
     }
 
     void BaseParser::error(const std::string& error, std::string exp)
@@ -164,22 +196,38 @@ namespace Ark::internal
         return false;
     }
 
-    bool BaseParser::comment()
+    bool BaseParser::comment(std::string* s)
     {
-        if (accept(IsChar('#')))
+        if (accept(IsChar('#'), s))
         {
-            while (accept(IsNot(IsChar('\n'))))
+            while (accept(IsNot(IsChar('\n')), s))
                 ;
-            accept(IsChar('\n'));
+            accept(IsChar('\n'), s);
             return true;
         }
         return false;
     }
 
-    bool BaseParser::newlineOrComment()
+    bool BaseParser::spaceComment(std::string* s)
     {
-        bool matched = space();
-        while (!isEOF() && comment())
+        bool matched = false;
+
+        inlineSpace();
+        while (!isEOF() && comment(s))
+        {
+            inlineSpace();
+            matched = true;
+        }
+
+        return matched;
+    }
+
+    bool BaseParser::newlineOrComment(std::string* s)
+    {
+        bool matched = false;
+
+        space();
+        while (!isEOF() && comment(s))
         {
             space();
             matched = true;
@@ -192,13 +240,11 @@ namespace Ark::internal
     {
         if (!accept(IsChar(c)))
             return false;
-        newlineOrComment();
         return true;
     }
 
     bool BaseParser::suffix(char c)
     {
-        newlineOrComment();
         return accept(IsChar(c));
     }