+++ src/arkreactor/Compiler/AST/Node.cpp
@@ -99,6 +99,22 @@ namespace Ark::internal
         m_filename = filename;
     }
 
+    Node& Node::attachNearestCommentBefore(const std::string& comment)
+    {
+        m_comment = comment;
+        return *this;
+    }
+
+    Node& Node::attachCommentAfter(const std::string& comment)
+    {
+        if (!m_after_comment.empty())
+            m_after_comment += "\n";
+        m_after_comment += comment;
+        if (!m_after_comment.empty() && m_after_comment.back() == '\n')
+            m_after_comment.pop_back();
+        return *this;
+    }
+
     std::size_t Node::line() const noexcept
     {
         return m_line;
@@ -112,6 +128,16 @@ namespace Ark::internal
     const std::string& Node::filename() const noexcept
     {
         return m_filename;
+    }
+
+    const std::string& Node::comment() const noexcept
+    {
+        return m_comment;
+    }
+
+    const std::string& Node::commentAfter() const noexcept
+    {
+        return m_after_comment;
     }
 
     std::string Node::repr() const noexcept
@@ -284,12 +310,8 @@ namespace Ark::internal
 
         if (A.m_type != NodeType::List)
             return A.m_value == B.m_value;
-
-        if (A.m_type == NodeType::List)
+        else
             throw TypeError("Can not compare lists");
-
-        // any other type => false (here, Closure)
-        return false;
     }
 
     bool operator<(const Node& A, const Node& B)
@@ -306,9 +328,6 @@ namespace Ark::internal
             case NodeType::String:
                 return A.m_value < B.m_value;
 
-            case NodeType::List:
-                // return A.m_list < B.m_list;  // fixme
-
             default:
                 return false;
         }
@@ -318,6 +337,8 @@ namespace Ark::internal
     {
         switch (A.nodeType())
         {
+            case NodeType::Field:
+                [[fallthrough]];
             case NodeType::List:
                 return A.constList().empty();
 
@@ -335,8 +356,6 @@ namespace Ark::internal
                 else if (A.string() == "false" || A.string() == "nil")
                     return true;
                 return false;
-
-                // todo: implement field?
 
             default:
                 return false;